新增：背包扩展到128格，理论上可以达到signed short极限

+++ src/main/java/tools/PacketCreator.java
@@ -501,7 +501,7 @@ public class PacketCreator {
 
     private static void addInventoryInfo(OutPacket p, Character chr) {
         for (byte i = 1; i <= 5; i++) {
-            p.writeByte(chr.getInventory(InventoryType.getByType(i)).getSlotLimit());
+            p.writeShort(chr.getInventory(InventoryType.getByType(i)).getSlotLimit());
         }
         p.writeLong(getTime(-2));
         Inventory iv = chr.getInventory(InventoryType.EQUIPPED);
